+++ src/main/java/roomescape/controller/ReservationController.java
@@ -1,4 +1,4 @@
-package roomescape.domain.reservation;
+package roomescape.controller;
 
 
 import lombok.RequiredArgsConstructor;
@@ -10,6 +10,7 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
+import roomescape.domain.reservation.ReservationService;
 import roomescape.dto.ReservationAddRequestDto;
 import roomescape.dto.ReservationResponseDto;
 import roomescape.domain.reservationtime.ReservationTimeService;
