+++ src/main/java/roomescape/controller/ThemeController.java
@@ -1,4 +1,4 @@
-package roomescape.domain.theme;
+package roomescape.controller;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
@@ -9,6 +9,7 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
+import roomescape.domain.theme.ThemeService;
 import roomescape.entities.Theme;
 import roomescape.dto.ThemeAddRequestDto;
